requirements.txt
asgiref==3.6.0
certifi==2022.12.7
distlib==0.3.6
Django==4.1.7
django-cors-headers==3.14.0
django-environ==0.10.0
djangorestframework==3.14.0
filelock==3.9.0
pipenv==2023.2.18
platformdirs==3.1.1
psycopg2==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
virtualenv==20.21.0
virtualenv-clone==0.5.7
